+++ src/store/modules/search.js
@@ -1,17 +1,22 @@
-import { search } from "@/api/search";
-import { isFileOwner, getFileOwnerName, getExternalFileOwnerName } from "@/utils/formatFile";
-import { getNetworkItemByAppId } from "@/utils/networkDest";
 import i18n from "@/i18n";
+import { advancedSearch } from "@/api/search";
+import { getNetworkItemByAppId } from "@/utils/networkDest";
+import { isFileOwner, getFileOwnerName, getExternalFileOwnerName } from "@/utils/formatFile";
 
 const actions = {
-  async fetchSearchFiles({ dispatch, commit }, query) {
+  async fetchSearchFiles({ dispatch, commit }, { query, pageNum, pageAmount, isAppend }) {
     try {
-      const results = await search(query);
-      commit("setFiles", results);
+      const tempResults = await advancedSearch(query, pageNum, pageAmount);
+      const results = tempResults.files.map((item) => item.file);
+
+      commit("setIsShared", true);
+      commit("updatePageNum", pageNum + 1);
+      isAppend ? commit("setAppendFiles", results) : commit("setFiles", results);
+      commit("setServerFilesLength", tempResults.count);
 
       results.forEach(async (file) => {
         if (file.isExternal) {
-          const networkDest = getNetworkItemByAppId(file.appId);
+          const networkDest = getNetworkItemByAppId(file.appID);
           const formattedFile = file;
           formattedFile.owner = await getExternalFileOwnerName(file.ownerId, networkDest.value);
           commit("updateFile", formattedFile);