+++ src/views/SharedWithMe.vue
@@ -8,9 +8,9 @@
 </template>
 
 <script>
-import { mapGetters } from "vuex";
 import PageTemplate from "@/components/BasePageTemplate";
-import { pageSize } from "@/config";
+import { mapGetters } from "vuex";
+import { getPageSize } from "@/utils/getPageSize";
 
 export default {
   name: "Shared",
@@ -24,30 +24,24 @@ export default {
         this.$store.commit("setFiles", []);
         this.$store.commit("setServerFilesLength", 1);
         this.$store.commit("updatePageNum", 1);
-        this.$store.dispatch("fetchSharedFiles", { pageNum: 0, pageAmount: this.getPageSize(newValue) });
+        this.$store.dispatch("fetchSharedFiles", { pageNum: 0, pageAmount: getPageSize(newValue) });
       }
     },
   },
   created() {
-    this.$store.dispatch("fetchSharedFiles", { pageNum: 0, pageAmount: this.getPageSize(this.fileView) });
+    this.$store.dispatch("fetchSharedFiles", { pageNum: 0, pageAmount: getPageSize(this.fileView) });
   },
   methods: {
-    getPageSize(fileView) {
-      return fileView === 0 || fileView === "undefined" || fileView === undefined ? pageSize : pageSize * 2;
-    },
     onPageChange(page) {
-      if (this.fileView == 1) {
-        if (this.getPageSize(this.fileView) * (page - 1) < this.serverFilesLength) {
-          this.$store.dispatch("fetchSharedFiles", {
-            pageNum: page - 1,
-            isAppend: true,
-            pageAmount: this.getPageSize(this.fileView),
-          });
-        }
-      } else {
-        this.$store.dispatch("fetchSharedFiles", {
+      if (
+        this.fileView != 1 ||
+        (this.fileView === 1 && getPageSize(this.fileView) * (page - 1) < this.serverFilesLength)
+      ) {
+        this.$store.dispatch("fetchSearchFiles", {
+          query: this.$route.query.q,
           pageNum: page - 1,
-          pageAmount: this.getPageSize(this.fileView),
+          isAppend: this.fileView === 1,
+          pageAmount: getPageSize(this.fileView),
         });
       }
     },